+++ frontend/src/routes/Search.tsx
@@ -4,7 +4,6 @@ import {
     LuCalendar,
     LuCalendarRange,
     LuLayout,
-    LuRadio,
     LuVolume2,
     LuX,
 } from "react-icons/lu";
@@ -43,15 +42,7 @@ import {
 import { RouterControl, makeRoute } from "../rauta";
 import { loadQuery } from "../relay";
 import { Link, useRouter } from "../router";
-import {
-    Creators,
-    Thumbnail,
-    ThumbnailImg,
-    ThumbnailOverlay,
-    ThumbnailOverlayContainer,
-    ThumbnailReplacement,
-    formatDuration,
-} from "../ui/Video";
+import { Creators, Thumbnail, formatDuration } from "../ui/Video";
 import { SmallDescription } from "../ui/metadata";
 import { Breadcrumbs, BreadcrumbsContainer, BreadcrumbSeparator } from "../ui/Breadcrumbs";
 import { MissingRealmName } from "./util";
@@ -69,6 +60,7 @@ import { DirectSeriesRoute, SeriesRoute } from "./Series";
 import { PartOfSeriesLink } from "../ui/Blocks/VideoList";
 import { SearchSlidePreviewQuery } from "./__generated__/SearchSlidePreviewQuery.graphql";
 import { RelativeDate } from "../ui/time";
+import { ThumbnailStack } from "../ui/Series";
 
 
 export const isSearchActive = (): boolean => document.location.pathname === "/~search";
@@ -423,7 +415,7 @@ const CenteredNote: React.FC<{ children: ReactNode }> = ({ children }) => (
 type Results = Extract<SearchQuery$data["search"], { __typename: "SearchResults" }>;
 type Item = Results["items"][number];
 type EventItem = Omit<Extract<Item, { "__typename": "SearchEvent" }>, "__typename">;
-type SeriesItem = Omit<Extract<Item, { "__typename": "SearchSeries" }>, "__typename">;
+export type SeriesItem = Omit<Extract<Item, { "__typename": "SearchSeries" }>, "__typename">;
 type RealmItem = Omit<Extract<Item, { "__typename": "SearchRealm" }>, "__typename">;
 
 type SearchResultsProps = {
@@ -846,12 +838,6 @@ const TextMatchTooltip: React.FC<TextMatchTooltipProps> = ({ previewImage, textM
 };
 
 
-type ThumbnailInfo = {
-    readonly audioOnly: boolean;
-    readonly live: boolean;
-    readonly url: string | null | undefined;
-}
-
 const SearchSeries: React.FC<SeriesItem> = ({
     id, title, description, thumbnailStack, matches, hostRealms,
 }) => {
@@ -894,131 +880,6 @@ const SearchSeries: React.FC<SeriesItem> = ({
     }}</Item>;
 };
 
-type ThumbnailStackProps = Pick<SeriesItem, "title" | "thumbnailStack">
-
-export const ThumbnailStack: React.FC<ThumbnailStackProps> = ({ thumbnailStack, title }) => {
-    const isDarkScheme = useColorScheme().scheme === "dark";
-
-    return (
-        <div css={{
-            zIndex: 0,
-            margin: "0 auto",
-            width: "70%",
-            display: "grid",
-            gridAutoColumns: "1fr",
-            "> div": {
-                position: "relative",
-                borderRadius: 8,
-                // The outline needs to be in a pseudo element as otherwise, it is
-                // hidden behind the img for some reason.
-                "::after": {
-                    content: "''",
-                    position: "absolute",
-                    inset: 0,
-                    borderRadius: 8,
-                    outline: `2px solid ${COLORS.neutral70}`,
-                    outlineOffset: -2,
-                },
-            },
-            "> div:not(:last-child)": {
-                boxShadow: "3px -2px 6px rgba(0, 0, 0, 40%)",
-            },
-            "> div:nth-child(1)": {
-                zIndex: 3,
-                gridColumn: "1 / span 10",
-                gridRow: "3 / span 10",
-            },
-            "> div:nth-child(2)": {
-                zIndex: 2,
-                gridColumn: "2 / span 10",
-                gridRow: "2 / span 10",
-            },
-            "> div:nth-child(3)": {
-                zIndex: 1,
-                gridColumn: "3 / span 10",
-                gridRow: "1 / span 10",
-            },
-        }}>
-            {thumbnailStack.thumbnails.slice(0, 3).map((info, idx) => <div key={idx}>
-                <SeriesThumbnail {...{ info, title }} />
-            </div>)}
-            {/* Add fake thumbnails to always have 3. The visual image of 3 things behind each other
-                is more important than actually showing the correct number of thumbnails. */}
-            {[...Array(Math.max(0, 3 - thumbnailStack.thumbnails.length))].map((_, idx) => (
-                <div key={"dummy" + idx}>
-                    <DummySeriesStackThumbnail isDark={isDarkScheme} />
-                </div>
-            ))}
-        </div>
-    );
-};
-
-const DummySeriesStackThumbnail: React.FC<{ isDark: boolean }> = ({ isDark }) => (
-    <ThumbnailOverlayContainer css={{
-        // Pattern from https://css-pattern.com/overlapping-cubes/,
-        // MIT licensed: https://github.com/Afif13/CSS-Pattern
-        "--s": "40px",
-        ...isDark ? {
-            "--c1": "#2c2c2c",
-            "--c2": "#292929",
-            "--c3": "#262626",
-        } : {
-            "--c1": "#e8e8e8",
-            "--c2": "#e3e3e3",
-            "--c3": "#dddddd",
-        },
-
-        "--_g": "0 120deg,#0000 0",
-        background: `
-            conic-gradient(             at calc(250%/3) calc(100%/3),
-                var(--c3) var(--_g)),
-            conic-gradient(from -120deg at calc( 50%/3) calc(100%/3),
-                var(--c2) var(--_g)),
-            conic-gradient(from  120deg at calc(100%/3) calc(250%/3),
-                var(--c1) var(--_g)),
-            conic-gradient(from  120deg at calc(200%/3) calc(250%/3),
-                var(--c1) var(--_g)),
-            conic-gradient(from -180deg at calc(100%/3) 50%,
-                var(--c2)  60deg,var(--c1) var(--_g)),
-            conic-gradient(from   60deg at calc(200%/3) 50%,
-                var(--c1)  60deg,var(--c3) var(--_g)),
-            conic-gradient(from  -60deg at 50% calc(100%/3),
-                var(--c1) 120deg,var(--c2) 0 240deg,var(--c3) 0)
-        `,
-        backgroundSize: "calc(var(--s)*sqrt(3)) var(--s)",
-    }} />
-);
-
-type SeriesThumbnailProps = {
-    info: ThumbnailInfo;
-    title: string;
-}
-
-const SeriesThumbnail: React.FC<SeriesThumbnailProps> = ({ info, title }) => {
-    const { t } = useTranslation();
-    const isDark = useColorScheme().scheme === "dark";
-
-    let inner;
-    if (info.url != null) {
-        // We have a proper thumbnail.
-        inner = <ThumbnailImg
-            src={info.url}
-            alt={t("series.entry-of-series-thumbnail", { series: title })}
-        />;
-    } else {
-        inner = <ThumbnailReplacement audioOnly={info.audioOnly} {...{ isDark }} />;
-    }
-
-    const overlay = <ThumbnailOverlay backgroundColor="rgba(200, 0, 0, 0.9)">
-        <LuRadio css={{ fontSize: 19, strokeWidth: 1.4 }} />
-        {t("video.live")}
-    </ThumbnailOverlay>;
-
-    return <ThumbnailOverlayContainer>
-        {inner}
-        {info.live && overlay}
-    </ThumbnailOverlayContainer>;
-};
 
 type SearchBreadcrumbsProps = {
     ancestorNames: readonly (string | null | undefined)[];
