+++ frontend/src/routes/manage/Series/SeriesDetails.tsx
@@ -1,14 +1,16 @@
+import { useTranslation } from "react-i18next";
 import { graphql, useMutation } from "react-relay";
 
 import i18n from "../../../i18n";
-import { makeManageSeriesRoute } from "./Shared";
-import { ManageSeriesRoute, SingleSeries } from ".";
+import { makeManageSeriesRoute, Series } from "./Shared";
+import { ManageSeriesRoute } from ".";
 import { DirectSeriesRoute } from "../../Series";
 import {
     UpdatedCreatedInfo,
     DetailsPage,
     DirectLink,
     MetadataSection,
+    DeleteButton,
 } from "../Shared/Details";
 import {
     SeriesDetailsMetadataMutation,
@@ -18,6 +20,12 @@ import {
 const updateSeriesMetadata = graphql`
     mutation SeriesDetailsMetadataMutation($id: ID!, $title: String!, $description: String) {
         updateSeriesMetadata(id: $id, title: $title, description: $description) { id }
+    }
+`;
+
+const deleteSeriesMutation = graphql`
+    mutation SeriesDetailsDeleteMutation($id: ID!) {
+        deleteSeries(id: $id) { id }
     }
 `;
 
@@ -32,19 +40,37 @@ export const ManageSeriesDetailsRoute = makeManageSeriesRoute(
             link: ManageSeriesRoute.url,
         }}
         sections={series => [
-            <UpdatedCreatedInfo key="created-info" item={series} />,
+            <UpdatedCreatedInfo key="date-info" item={series} />,
+            <SeriesButtonSection key="button-section" seriesId={series.id} />,
             <DirectLink key="direct-link" url={
                 new URL(DirectSeriesRoute.url({ seriesId: series.id }), document.baseURI)
             } />,
-            <SeriesMetadataSection
-                key="metadata"
-                series={series}
-            />,
+            <SeriesMetadataSection key="metadata" series={series} />,
         ]}
     />,
 );
 
-const SeriesMetadataSection: React.FC<{ series: SingleSeries }> = ({ series }) => {
+const SeriesButtonSection: React.FC<{ seriesId: string }> = ({ seriesId }) => {
+    const { t } = useTranslation();
+    const [commit] = useMutation(deleteSeriesMutation);
+
+    return <div css={{ display: "flex", gap: 12, marginBottom: 16 }}>
+        <DeleteButton
+            itemId={seriesId}
+            itemType="series"
+            returnPath="/~manage/series"
+            commit={config => {
+                const disposable = commit(config);
+                return { [Symbol.dispose]: () => disposable.dispose() };
+            }}
+        >
+            <br />
+            <p>{t("manage.my-series.delete-note")}</p>
+        </DeleteButton>
+    </div>;
+};
+
+const SeriesMetadataSection: React.FC<{ series: Series }> = ({ series }) => {
     const [commit, inFlight]
         = useMutation<SeriesDetailsMetadataMutation>(updateSeriesMetadata);
