+++ frontend/src/routes/manage/Video/index.tsx
@@ -1,6 +1,5 @@
-import { useTranslation } from "react-i18next";
 import { graphql } from "react-relay";
-import { match, useColorScheme } from "@opencast/appkit";
+import { match } from "@opencast/appkit";
 
 import { ManageNav } from "..";
 import { RootLoader } from "../../../layout/Root";
@@ -11,24 +10,15 @@ import {
 } from "./__generated__/VideoManageQuery.graphql";
 import { makeRoute } from "../../../rauta";
 import { loadQuery } from "../../../relay";
-import { Link } from "../../../router";
 import { NotAuthorized } from "../../../ui/error";
 import { Thumbnail } from "../../../ui/Video";
 import { keyOfId } from "../../../util";
-import { SmallDescription } from "../../../ui/metadata";
-import { COLORS } from "../../../color";
-import { InfoWithTooltip } from "../../../ui";
-import { relativeDate } from "../../../ui/time";
-import CONFIG from "../../../config";
 import {
     ColumnProps,
     createQueryParamsParser,
     DateColumn,
-    descriptionStyle,
     ManageItems,
     TableRow,
-    thumbnailLinkStyle,
-    titleLinkStyle,
 } from "../Shared/Table";
 
 
@@ -120,79 +110,14 @@ export const videoColumns: ColumnProps<Event>[] = [
     },
 ];
 
-export const EventRow: React.FC<{ item: Event }> = ({ item }) => {
-    const link = `${PATH}/${keyOfId(item.id)}`;
+export const EventRow: React.FC<{ item: Event }> = ({ item }) => <TableRow
+    itemType="video"
+    item={item}
+    link={`${PATH}/${keyOfId(item.id)}`}
+    thumbnail={deletionIsPending => <Thumbnail event={item} {...{ deletionIsPending }} />}
+    customColumns={videoColumns.map(col => <col.column key={col.key} item={item} />)}
+/>;
 
-    const deletionIsPending = Boolean(item.tobiraDeletionTimestamp);
-    const deletionDate = new Date(item.tobiraDeletionTimestamp ?? "");
-
-    // This checks if the current time is later than the deletion timestamp + twice
-    // the configured poll period to ensure at least one sync has taken place
-    // (+ 1min to allow some time for the Opencast delete job).
-    // If it is, the deletion in Opencast has possibly failed.
-    const pollPeriod = CONFIG.sync.pollPeriod * 1000;
-    const deletionFailed = Boolean(item.tobiraDeletionTimestamp
-        && Date.parse(item.tobiraDeletionTimestamp) + pollPeriod * 2 + 60000 < Date.now());
-
-    return <TableRow
-        thumbnail={deletionIsPending
-            ? <Thumbnail event={item} deletionIsPending={deletionIsPending} />
-            : <Link to={link} css={{ ...thumbnailLinkStyle }}>
-                <Thumbnail event={item} />
-            </Link>
-        }
-        title={deletionIsPending
-            ? <span css={{ color: COLORS.neutral60 }}>{item.title}</span>
-            : <Link to={link} css={{ ...titleLinkStyle }}>{item.title}</Link>
-        }
-        description={deletionIsPending
-            ? <PendingDeletionBody {...{ deletionFailed, deletionDate }} />
-            : <SmallDescription css={{ ...descriptionStyle }} text={item.description} />
-        }
-        syncInfo={{
-            isSynced: !!item.syncedData,
-            notReadyLabel: "video.not-ready.label",
-        }}
-        customColumns={videoColumns.map(col => <col.column key={col.key} item={item} />)}
-    />;
-};
-
-type PendingDeleteBodyProps = {
-    deletionFailed: boolean;
-    deletionDate: Date;
-}
-
-const PendingDeletionBody: React.FC<PendingDeleteBodyProps> = ({
-    deletionFailed, deletionDate,
-}) => {
-    const isDark = useColorScheme().scheme === "dark";
-    const { t } = useTranslation();
-
-    const now = Date.now();
-    const [, relative] = relativeDate(deletionDate, now);
-
-    return (
-        <div css={{
-            color: isDark ? COLORS.neutral60 : COLORS.neutral50,
-            display: "flex",
-            fontSize: 13,
-            marginTop: 4,
-            padding: "0 4px",
-        }}>
-            <span css={{ fontStyle: "italic" }}>
-                {t(`manage.my-videos.details.delete.${
-                    deletionFailed ? "failed-maybe" : "pending"
-                }`)}
-            </span>
-            <InfoWithTooltip
-                tooltip={t(`manage.my-videos.details.delete.tooltip.${
-                    deletionFailed ? "failed" : "pending"
-                }`, { time: relative })}
-                mode={deletionFailed ? "warning" : "info"}
-            />
-        </div>
-    );
-};
 
 const parseVideosColumn = (sortBy: string | null): VideosSortColumn =>
     sortBy !== null
