+++ frontend/src/ui/Video.tsx
@@ -153,9 +153,10 @@ export const BaseThumbnailReplacement: React.FC<BaseThumbnailReplacementProps> =
 type ThumbnailOverlayProps = PropsWithChildren<{
     backgroundColor: string;
 }>;
-export const ThumbnailOverlay: React.FC<ThumbnailOverlayProps> = (
-    { children, backgroundColor }
-) => (
+export const ThumbnailOverlay: React.FC<ThumbnailOverlayProps> = ({
+    children,
+    backgroundColor,
+}) => (
     <div css={{
         display: "inline-flex",
         alignItems: "center",
@@ -173,10 +174,10 @@ export const ThumbnailOverlay: React.FC<ThumbnailOverlayProps> = (
     </div>
 );
 
-type ThumbnailOverlayContainerProps = JSX.IntrinsicElements["div"];
-export const ThumbnailOverlayContainer: React.FC<ThumbnailOverlayContainerProps> = (
-    { children, ...rest }
-) => {
+export const ThumbnailOverlayContainer: React.FC<JSX.IntrinsicElements["div"]> = ({
+    children,
+    ...rest
+}) => {
     const isDark = useColorScheme().scheme === "dark";
 
     return <div css={{
